backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
crewai>=1.9.0
crewai-tools>=0.17.0
chromadb>=0.4.0
supabase>=2.11.0
python-dotenv>=1.0.1
pydantic-settings>=2.7.0
anthropic>=0.43.0
